Add importing a minimal JWK
As stated in section 6.3.2 of RFC 7518 (JWA), JWK private keys
containing only the n, e and d parameters are valid.

This make sure these keys can be imported by finding the p and q factors
when loading the JWK parameters, just as is already done with the dp, dq
and qi parameters.

Closes #158

Signed-off-by: Jacob Kiers <kiers@comandi.nl>